Add files via upload
First version of the code. Still quite incomplete (half of the log-linearized equations written, but still all steady state equations to write). Code is written to use a function like gensys in matlab to solve the model (if no such function exists it is probably simple to use the matlab one and change it a bit to work in Julia).