Add files via upload
Corrected a few things in the code. Changed things in the gensys part of the code and it is now working. The code is now able to take parameters, solve the model and give us the state-space matrices that one can use to simulate the model or that will be needed for estimation later.