Add files via upload
Reuploading the previous file mycode.jl now with the name model.jl. 

Made some changes in the structure: The file model.jl now has a bunch of functions necessary to solve the model, which are called in the file main.jl.
The variablesindices.jl relates variables to indices in the state space and is in a separate jl file because it might have to be called in different places.