Add files via upload
Changed a small thing in the multiplier. Developed a function for the prior predictive analysis, and one function that imposes the model restrictions. I have been able to get some of the values from column 1 of Table 3, for the probabilities that the multiplier is above one on impact. However, the value for the case with 4 quarters seems not to be correct. I will look into this (there are two possibilities: 1) there is something wrong in the transition matrix, so that impact multipliers are correct but not the rest; 2) the computation of the multiplier itself has some mistake for the case of t>1.